Refs #28643 -- Moved db function mixins to a separate module.
